- category api 적용 완료

+++ app/src/main/java/com/dod/whateat/service/FoodService.kt
@@ -8,7 +8,7 @@ class FoodService {
     fun randomFoodList(): MutableList<FoodData> {
         val list = mutableListOf<FoodData>()
         for(i in 0..100){
-            list.add(FoodData(i, "$i 번째 food", 0))
+            list.add(FoodData(i.toLong(), "$i 번째 food", 0))
         }
         return list
     }
@@ -16,7 +16,7 @@ class FoodService {
     fun foodList(categorySeq: Int): DefaultData {
         val list = mutableListOf<FoodData>()
         for(i in 0..100){
-            list.add(FoodData(i, "$i 번째 food", categorySeq))
+            list.add(FoodData(i.toLong(), "$i 번째 food", categorySeq))
         }
 
         return DefaultData(1, Gson().toJsonTree(list))
